third_party/WebKit/LayoutTests/media/track/opera/interfaces/TextTrackCueList/getter-expected.txt

 
PASS TextTrackCueList getter 
FAIL TextTrackCueList getter, no indexed set/create assert_equals: expected undefined but got "foo"
FAIL TextTrackCueList getter, no indexed set/create (strict) assert_equals: expected undefined but got "foo"
